+++ src/lib.rs
@@ -2,7 +2,7 @@ mod error;
 pub mod models;
 mod whatspapp_client;
 
-pub use crate::whatspapp_client::WhatasppClient;
+pub use crate::whatspapp_client::WhatsappClient;
 pub use error::WhatsappError;
 
 pub const WHATSAPP: &str = "whatsapp";